make util file for methods shared by plain text and tabix gtf adapters
