+++ plugins/gtf/src/index.ts
@@ -2,6 +2,7 @@ import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
 import PluginManager from '@jbrowse/core/PluginManager'
 import Plugin from '@jbrowse/core/Plugin'
 import { configSchema as gtfTabixAdapterConfigSchema } from './GtfTabixAdapter'
+import { configSchema as gtfAdapterConfigSchema } from './GtfAdapter'
 import { FileLocation } from '@jbrowse/core/util/types'
 import {
   makeIndex,
@@ -43,6 +44,36 @@ export default class GtfPlugin extends Plugin {
                 location: index || makeIndex(file, '.tbi'),
                 indexType: makeIndexType(indexName, 'CSI', 'TBI'),
               },
+            }
+          }
+          return adapterGuesser(file, index, adapterHint)
+        }
+      },
+    )
+    pluginManager.addAdapterType(
+      () =>
+        new AdapterType({
+          name: 'GtfAdapter',
+          configSchema: gtfAdapterConfigSchema,
+          getAdapterClass: () =>
+            import('./GtfAdapter/GtfAdapter').then(r => r.default),
+        }),
+    )
+    pluginManager.addToExtensionPoint(
+      'Core-guessAdapterForLocation',
+      (adapterGuesser: AdapterGuesser) => {
+        return (
+          file: FileLocation,
+          index?: FileLocation,
+          adapterHint?: string,
+        ) => {
+          const regexGuess = /\.gtf?$/i
+          const adapterName = 'GtfAdapter'
+          const fileName = getFileName(file)
+          if (regexGuess.test(fileName) || adapterHint === adapterName) {
+            return {
+              type: adapterName,
+              gtfLocation: file,
             }
           }
           return adapterGuesser(file, index, adapterHint)
